Update to user management tests with Okta enabled / disabled
